Brunch 에서 Branch로 이름 번경, 글 버튼 삭제, 강요한으로 이름 번경 및 작가 소개 변경, 작품UI 번경, 글 삭제 기능 삭제
