+++ Assets/02.Scripts/SO/BranchPostDataSO.cs
@@ -2,12 +2,12 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu(menuName = "SO/Site/Brunch/BrunchWriteData")]
-public class BrunchPostDataSO : ScriptableObject
+[CreateAssetMenu(menuName = "SO/Site/Branch/BranchWriteData")]
+public class BranchPostDataSO : ScriptableObject
 {
     public string wirteTitle;
     public string wirteInfo;
     public string wirteDate;
-    public EWorkKeys workTitle;
+    public EBranchWorkCategory workTitle;
     public Sprite writeImage;
 }
