+++ Assets/02.Scripts/UI/Browser/Site/Sites/Brunch/BranchPostRemoveButton.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e1b0bf6d057d4700b5ac1e023ce9c9cf
+guid: 451d20341415a784298e53da782ec668
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
